python/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bitarray==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.7
ckzg==2.1.7
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==5.3.1
eth_abi==6.0.0b1
exceptiongroup==1.3.1
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
hyperliquid-python-sdk==0.22.0
idna==3.11
msgpack==1.1.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
parsimonious==0.10.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
rlp==4.1.0
six==1.17.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websocket-client==1.9.0
websockets==16.0
